third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-mixed-content-to-inscope.html

<!DOCTYPE html>
<title>Service Worker: Mixed content of fetch()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/get-host-info.js?pipe=sub"></script>
<script src="resources/test-helpers.js"></script>
<body></body>
<script>
async_test(function(t) {
    var host_info = get_host_info();
    window.addEventListener('message', t.step_func(on_message), false);
    with_iframe(
      host_info['HTTPS_ORIGIN'] + base_path() +
      'resources/fetch-mixed-content-iframe.html?target=inscope');
    function on_message(e) {
      assert_equals(e.data.results, 'finish');
      t.done();
    }
  }, 'Verify Mixed content of fetch() in a Service Worker');
</script>
